+++ wow_dbc/CHANGELOG.md
@@ -26,6 +26,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 * BREAKING: `Language` renamed to `ClientLanguage`.
 * BREAKING: `Category` renamed to `ServerCategory`.
 * BREAKING: `Type` renamed to `OceanType`.
+* BREAKING: `ReputationRaceMask` renamed to `AllowedRace`.
 
 ### Removed